tried using path from config file, faced errors, so reverted back
