+++ package.json
@@ -68,7 +68,7 @@
         "ts-node": "ts-node",
         "tsc": "tsc",
         "test": "echo \"Error: no test specified\" && exit 1",
-        "serve": "nodemon -r dotenv/config server.js"
+        "serve": "ts-node --esm -r dotenv/config ./server.ts"
     },
     "type": "module"
 }
